separate logic and presentation for event website block
